refactor: Phase 4 - split models.py into models/ (5 domain files)
- models/narration.py: NarrationStyle, EmotionType, NarrationBlock
- models/video.py: TimeRange, VideoSegment, EmotionPeak
- models/media.py: SubtitleItem, AudioTrack
- models/project.py: VideoProject, VideoGroup, TaskProgress
- models/__init__.py: re-exports from domain files
- models.py: compat layer with DeprecationWarning

Test: 384 passed, 20 skipped
